refactor: move validation logic to validators
