+++ src/adapters/chartOfAccount.adapter.ts
@@ -33,9 +33,6 @@ export class ChartOfAccountAdapter extends BaseAdapter<ChartOfAccount> {
   ];
 
   protected override async onBeforeCreate(data: Partial<ChartOfAccount>): Promise<Partial<ChartOfAccount>> {
-    // Validate account data
-    this.validateAccountData(data);
-    
     // Set default values
     if (data.is_active === undefined) {
       data.is_active = true;
@@ -50,11 +47,7 @@ export class ChartOfAccountAdapter extends BaseAdapter<ChartOfAccount> {
   }
 
   protected override async onBeforeUpdate(id: string, data: Partial<ChartOfAccount>): Promise<Partial<ChartOfAccount>> {
-    // Validate account data if fields are being updated
-    if (data.code || data.name || data.account_type) {
-      this.validateAccountData(data);
-    }
-    
+    // Repositories handle validation
     return data;
   }
 
@@ -94,22 +87,6 @@ export class ChartOfAccountAdapter extends BaseAdapter<ChartOfAccount> {
     await logAuditEvent('delete', 'chart_of_account', id, { id });
   }
 
-  private validateAccountData(data: Partial<ChartOfAccount>): void {
-    if (data.code !== undefined && !data.code.trim()) {
-      throw new Error('Account code is required');
-    }
-    
-    if (data.name !== undefined && !data.name.trim()) {
-      throw new Error('Account name is required');
-    }
-    
-    if (data.account_type !== undefined) {
-      const validTypes = ['asset', 'liability', 'equity', 'revenue', 'expense'];
-      if (!validTypes.includes(data.account_type)) {
-        throw new Error('Invalid account type. Must be one of: asset, liability, equity, revenue, expense');
-      }
-    }
-  }
 
   public async getHierarchy(): Promise<ChartOfAccount[]> {
     const { data, error } = await supabase.rpc('get_chart_of_accounts_hierarchy');
