+++ src/repositories/chartOfAccount.repository.ts
@@ -3,6 +3,7 @@ import { BaseRepository } from './base.repository';
 import { ChartOfAccount } from '../models/chartOfAccount.model';
 import { ChartOfAccountAdapter } from '../adapters/chartOfAccount.adapter';
 import { NotificationService } from '../services/NotificationService';
+import { ChartOfAccountValidator } from '../validators/chartOfAccount.validator';
 
 @injectable()
 export class ChartOfAccountRepository extends BaseRepository<ChartOfAccount> {
@@ -11,8 +12,8 @@ export class ChartOfAccountRepository extends BaseRepository<ChartOfAccount> {
   }
 
   protected override async beforeCreate(data: Partial<ChartOfAccount>): Promise<Partial<ChartOfAccount>> {
-    // Additional repository-level validation
-    this.validateAccountData(data);
+    // Validate account data
+    ChartOfAccountValidator.validate(data);
     
     // Format data before creation
     return this.formatAccountData(data);
@@ -24,8 +25,8 @@ export class ChartOfAccountRepository extends BaseRepository<ChartOfAccount> {
   }
 
   protected override async beforeUpdate(id: string, data: Partial<ChartOfAccount>): Promise<Partial<ChartOfAccount>> {
-    // Additional repository-level validation
-    this.validateAccountData(data);
+    // Validate account data
+    ChartOfAccountValidator.validate(data);
     
     // Format data before update
     return this.formatAccountData(data);
@@ -50,32 +51,6 @@ export class ChartOfAccountRepository extends BaseRepository<ChartOfAccount> {
   }
 
   // Private helper methods
-  private validateAccountData(data: Partial<ChartOfAccount>): void {
-    const errors: string[] = [];
-
-    // Basic validation
-    if (data.code !== undefined && !data.code.trim()) {
-      errors.push('Account code is required');
-    }
-    
-    if (data.name !== undefined && !data.name.trim()) {
-      errors.push('Account name is required');
-    }
-    
-    if (data.account_type !== undefined) {
-      const validTypes = ['asset', 'liability', 'equity', 'revenue', 'expense'];
-      if (!validTypes.includes(data.account_type)) {
-        errors.push('Invalid account type. Must be one of: asset, liability, equity, revenue, expense');
-      }
-    }
-
-    if (errors.length > 0) {
-      errors.forEach(error => {
-        NotificationService.showError(error, 5000);
-      });
-      throw new Error('Validation failed: ' + errors.join(', '));
-    }
-  }
 
   private formatAccountData(data: Partial<ChartOfAccount>): Partial<ChartOfAccount> {
     const formattedData = { ...data };
