+++ src/repositories/member.repository.ts
@@ -3,6 +3,7 @@ import { BaseRepository } from './base.repository';
 import { Member } from '../models/member.model';
 import { MemberAdapter } from '../adapters/member.adapter';
 import { NotificationService } from '../services/NotificationService';
+import { MemberValidator } from '../validators/member.validator';
 
 @injectable()
 export class MemberRepository extends BaseRepository<Member> {
@@ -11,8 +12,8 @@ export class MemberRepository extends BaseRepository<Member> {
   }
 
   protected override async beforeCreate(data: Partial<Member>): Promise<Partial<Member>> {
-    // Additional repository-level validation
-    this.validateMemberData(data);
+    // Validate member data
+    MemberValidator.validate(data);
     
     // Format data before creation
     return this.formatMemberData(data);
@@ -24,8 +25,8 @@ export class MemberRepository extends BaseRepository<Member> {
   }
 
   protected override async beforeUpdate(id: string, data: Partial<Member>): Promise<Partial<Member>> {
-    // Additional repository-level validation
-    this.validateMemberData(data);
+    // Validate member data
+    MemberValidator.validate(data);
     
     // Format data before update
     return this.formatMemberData(data);
@@ -47,41 +48,6 @@ export class MemberRepository extends BaseRepository<Member> {
   }
 
   // Private helper methods
-  private validateMemberData(data: Partial<Member>): void {
-    const errors: string[] = [];
-
-    // Basic Info validation
-    if (!data.first_name?.trim()) {
-      errors.push('First name is required');
-    }
-    if (!data.last_name?.trim()) {
-      errors.push('Last name is required');
-    }
-    if (!data.gender) {
-      errors.push('Gender is required');
-    }
-    if (!data.marital_status) {
-      errors.push('Marital status is required');
-    }
-
-    // Contact Info validation
-    if (!data.contact_number?.trim()) {
-      errors.push('Contact number is required');
-    }
-    if (!data.address?.trim()) {
-      errors.push('Address is required');
-    }
-    if (data.email && !/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(data.email)) {
-      errors.push('Invalid email format');
-    }
-
-    if (errors.length > 0) {
-      errors.forEach(error => {
-        NotificationService.showError(error, 5000);
-      });
-      throw new Error('Validation failed: ' + errors.join(', '));
-    }
-  }
 
   private formatMemberData(data: Partial<Member>): Partial<Member> {
     return {
